Folder rename approach build failed fix
